feat: make error message display only when needed
This commit makes the error message display only when it is
required.That is, it displays the error message only when the user
is trying to log in. The error message will be removed if it was failed
log in attempt and the user is in the process of retyping.